Implemented common styles to the following connectors: 1. Amazon History 2. Chrome History 3. Firefox History 4. Instagram 5. LinkedIn 6. Twitter 7. Foursquare
These are all of the connectors that have clients in javascript.
I will not be able to implement something similar to the other, since they are written in Python and I have no knowledge of it.

Note:
All of the colors are the same in all clients this can be easily changed/modified by changing two elements of the array parameter of formatHTML.
First element is the color of the background of the header, and the second element is the color of the text in the header.