A dirt simple journal connector and an app that uses it
SimpleJournal is a connector that is as basic as it can get.  You can
put and then get data out of it.  Future enhancement is edit and delete.
All entries are timestampped, other than that they are just json
objects.

BloodPressureTracker is a basic app that I use on the SimpleJournal to
track my blood pressure daily.  It graphs it out and has a fairly
complete TODO list to grow from.  Might eventually merge into a general
health app.