New buttons on You-connect; fix after-effects
This covers most of #630.

The shared-ui submodule now contains .action-button for use when we want a nice
button.

I also noticed that there were supposed to be some after-effects of connecting
to a service that didn't appear to be working. Now the newly-connected service
icon disappears from the "More Services" section and appears in "Connected".

Also cleaned up the vertical alignment in the services list and did a bit of
linting.

Future work:
The Connect vs Connected text makes the button wider upon change. We could make
all the buttons the same width, or shrink the font on Connected. Not worrying
about it for now while I propagate the standard button first.