Convert the Got It button to CSS
More on #630.

This button is on a modal that appears your first time logging in after initial
setup. It's the last thing that appears to have an image background.

I know we want this style on more buttons, but I cannot for the life of me find
them in the UI. That seems like a problem.