The locker will now try like hell to find an avatar.
If you provide one through the account settings page, it will always use
that one. Then, it will traverse the map, trying the following in order:

* Twitter
* Facebook
* Github
* Foursquare
* Instagram
* last.fm

If none of the above are available, it will fall back to the old default
image, but that should now happen fairly rarely.

All of the logic is in a single function, so it can probably be grabbed
and transplanted into the Profiles collection whenever that gets
created.