Mostly fix Back/Forward on the Develop tab
The links inside the develop iframe were replacing the entire window URL,
causing jQuery.history to start a new stack. Now we only change the hash if
possible, and when we have an external link, we replace the whole URL.

This fixes #750, but I'm going to open a new bug to handle the very last page
you were on. If you visit Develop tabs A, B and C, then go somewhere external,
when you hit back you land on B, not C, because jQuery.history didn't see C
happen. To fix this, we could open external links in new windows, or manually
push C onto history's stack.