Open external links in new tab; simplify /develop
Closes #773

The /develop navigation's JS handler doesn't care if you were middle-clicking,
etc., which made people sad trying to open things in new windows. We'll probably
want to make that work better in general, but for now the external links will
always open in a new window, which might be good enough.

While I was in there, I simplified some of the HTML/CSS and collapsed a big case
statement responsible for making the nav links look selected.