Tighten up first-login experience
Fixes #764

The Explore-connect page was getting into various states of mixed element
visibility. Now, both nav sections are entirely hidden until you auth something
besides Github.

The connector list was also showing everything immediately, but we want to limit
it to Facebook or Twitter at first because those are the best sources or a
variety of data. Once you auth one of those, you'll get the full list.

Tangentially, the "start exploring" link was pointing to a bad URL. Now it just
goes to /explore.

Bonus:
- Refactored much of connect.js
- New whitespace standards
- Broke up some long lines
- Linted the JS files