Notify Airbrake of 500s when configured
Express has its own error handling that masks attempts to catch them from an
`uncaughtException` handler in `lockerd`. Now if an Airbrake API key is
available in the configs, Airbrake will be initialized on the locker and
notified in Express's `error` handler.

For testing, you can hit `/core/error` to trigger a 500.