Don't save the currently playing track.
When the scrobbles synclet runs, it needs to generate a key for each
play. It uses the timestamp, as that's guaranteed (by the rules of
scrobbling) to be distinct for each track. However, when the synclet
fires while the locker owner is listening to music through a service
that scrobbles, it pushes the current track on to the head of the
scrobble list with the attribute nowPlaying and no timestamp.

The synclet manager chokes if it gets entities without timestamps, and
it doesn't make any sense to save an ephemeral record anyway, so don't
save tracks that are currently playing or otherwise missing a timestamp.